+++ microservices/idam/identity-session-service/impl/src/controllers/auth_refresh.rs
@@ -50,7 +50,7 @@ pub fn handle(req: TypedHandlerRequest<Request>) -> Response {
         mfa_required: None,
         phone_verified: None,
         refresh_token: "refreshed-refresh".to_string(),
-        refresh_token_expires_in: Some(refresh_ttl_secs as i64),
+        refresh_token_expires_in: Some(refresh_ttl_secs as i32),
         scope: None,
         token_type: "Bearer".to_string(),
         user_id: "default".to_string(),
